+++ database/migrations/2024_12_31_025022_create_sales_details_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('quantity');
             $table->bigInteger('price');
             $table->bigInteger('total');
-            $table->bigInteger('diskon_barang')->nullable(); // Kolom untuk menyimpan diskon
+            $table->decimal('diskon_barang', 10, 0)->nullable(); // Kolom untuk menyimpan diskon
             $table->timestamps();
         });
     }
